gcp-cloud-functions-scripts/gcp-mwave/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery
pandas
datetime
bs4
requests